requirements2.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.0
pandas==2.1.0
pyarrow==14.0.0